requirements.txt
torch==2.3.1
transformers==4.41.2
datasets==2.20.0
evaluate==0.4.2
accelerate==0.33.0

sentence-transformers==2.7.0
faiss-cpu==1.8.0.post1
pypdf==4.3.1

scikit-learn==1.5.1
numpy==1.26.4
pandas==2.2.2

fastapi==0.112.1
uvicorn==0.30.6
gradio==4.44.1
pydantic==2.9.2
tqdm==4.66.4
